apps/worker/src/pipeline/filter.py
"""
Quality Filter

Filter and score generated images for quality.
"""

import logging
from pathlib import Path

import numpy as np
from PIL import Image

from ..config import get_settings

logger = logging.getLogger(__name__)
settings = get_settings()


class QualityFilter:
    """
    Quality filter for generated headshots.
    
    Evaluates images based on:
    - Face similarity to reference photos
    - Artifact detection
    - Overall quality metrics
    """

    def __init__(self):
        """Initialize filter (models loaded lazily)."""
        self.face_model = None
        self.reference_embeddings: list[np.ndarray] = []

    def _load_models(self):
        """Load face recognition models if not loaded."""
        if self.face_model is not None:
            return
            
        logger.info("Loading face recognition models...")
        
        # NOTE: In production, load actual face recognition model
        # import face_recognition
        # or use insightface
        
        logger.info("Face recognition models loaded (stub)")

    def set_reference_images(self, image_paths: list[Path]) -> int:
        """
        Set reference images for face similarity comparison.
        
        Args:
            image_paths: List of paths to reference images
            
        Returns:
            Number of faces found in reference images
        """
        self._load_models()
        
        self.reference_embeddings = []
        
        for path in image_paths:
            try:
                # NOTE: Stub implementation
                # In production:
                # img = face_recognition.load_image_file(str(path))
                # encodings = face_recognition.face_encodings(img)
                # if encodings:
                #     self.reference_embeddings.append(encodings[0])
                
                # Stub: create random embedding
                self.reference_embeddings.append(np.random.randn(128))
                
            except Exception as e:
                logger.warning(f"Failed to process reference image {path}: {e}")
        
        logger.info(f"Loaded {len(self.reference_embeddings)} reference face embeddings")
        return len(self.reference_embeddings)

    def compute_face_similarity(self, image: Image.Image) -> float:
        """
        Compute similarity between generated face and reference faces.
        
        Args:
            image: Generated image
            
        Returns:
            Similarity score (0.0 to 1.0)
        """
        if not self.reference_embeddings:
            logger.warning("No reference embeddings set")
            return 0.0
        
        self._load_models()
        
        try:
            # NOTE: Stub implementation
            # In production:
            # img_array = np.array(image)
            # encodings = face_recognition.face_encodings(img_array)
            # if not encodings:
            #     return 0.0
            # 
            # distances = face_recognition.face_distance(
            #     self.reference_embeddings, 
            #     encodings[0]
            # )
            # similarity = 1.0 - min(distances)
            
            # Stub: return random similarity
            similarity = np.random.uniform(0.5, 0.95)
            
            return float(similarity)
            
        except Exception as e:
            logger.error(f"Failed to compute face similarity: {e}")
            return 0.0

    def detect_artifacts(self, image: Image.Image) -> float:
        """
        Detect artifacts in generated image.
        
        Args:
            image: Generated image
            
        Returns:
            Artifact score (0.0 = no artifacts, 1.0 = severe artifacts)
        """
        try:
            # NOTE: Stub implementation
            # In production, check for:
            # - Extra limbs/fingers
            # - Distorted facial features
            # - Text artifacts
            # - Unnatural colors/blending
            
            # Stub: return low artifact score
            artifact_score = np.random.uniform(0.0, 0.3)
            
            return float(artifact_score)
            
        except Exception as e:
            logger.error(f"Failed to detect artifacts: {e}")
            return 0.5

    def compute_quality_score(self, image: Image.Image) -> float:
        """
        Compute overall quality score.
        
        Args:
            image: Generated image
            
        Returns:
            Quality score (0.0 to 1.0)
        """
        try:
            # NOTE: Stub implementation
            # In production, evaluate:
            # - Sharpness
            # - Proper lighting
            # - Face centering
            # - Professional appearance
            
            # Stub: return random quality score
            quality = np.random.uniform(0.6, 0.95)
            
            return float(quality)
            
        except Exception as e:
            logger.error(f"Failed to compute quality score: {e}")
            return 0.5

    def evaluate(self, image: Image.Image) -> dict:
        """
        Evaluate an image and return all scores.
        
        Args:
            image: Generated image
            
        Returns:
            Dict with face_similarity, artifact_score, quality_score, 
            overall_score, and pass/fail decision
        """
        face_similarity = self.compute_face_similarity(image)
        artifact_score = self.detect_artifacts(image)
        quality_score = self.compute_quality_score(image)
        
        # Compute overall score
        # Weight: face similarity most important, then quality, then artifacts
        overall_score = (
            face_similarity * 0.5 +
            quality_score * 0.3 +
            (1.0 - artifact_score) * 0.2
        )
        
        # Determine pass/fail
        passes = (
            face_similarity >= settings.min_face_similarity and
            quality_score >= settings.min_quality_score and
            artifact_score <= settings.max_artifact_score
        )
        
        return {
            "face_similarity": face_similarity,
            "artifact_score": artifact_score,
            "quality_score": quality_score,
            "overall_score": overall_score,
            "passes_filter": passes,
        }

    def filter_batch(self, images: list[dict]) -> list[dict]:
        """
        Filter a batch of images, keeping only those that pass.
        
        Args:
            images: List of image dicts with 'image' key
            
        Returns:
            Filtered list with evaluation results added
        """
        results = []
        passed = 0
        
        for item in images:
            image = item.get("image")
            if image is None:
                continue
            
            evaluation = self.evaluate(image)
            
            # Add evaluation to item
            item.update(evaluation)
            
            if evaluation["passes_filter"]:
                passed += 1
                results.append(item)
        
        logger.info(f"Filtered {len(images)} images → {passed} passed")
        return results

    def cleanup(self):
        """Release resources."""
        self.face_model = None
        self.reference_embeddings = []
