apps/worker/requirements.txt
# PyTorch (MPS-compatible)
torch>=2.1.0
torchvision>=0.16.0

# Diffusers and transformers
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.0

# LoRA training
peft>=0.7.0

# Image processing
Pillow>=10.2.0
opencv-python>=4.9.0

# Face detection/recognition
face-recognition>=1.3.0
dlib>=19.24.0
insightface>=0.7.3
onnxruntime>=1.16.0

# S3/MinIO
boto3>=1.34.0
botocore>=1.34.0

# Redis
redis>=5.0.0

# HTTP client (for API calls)
httpx>=0.26.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Progress and logging
tqdm>=4.66.0
rich>=13.7.0

# Development
pytest>=8.0.0
ruff>=0.2.0
